+++ subprojects/crengine/crengine/src/epubfmt.cpp
@@ -112,7 +112,7 @@ void ReadEpubNcxToc( ldomDocument * doc, ldomNode * mapRoot, LVTocItem * baseToc
     lUInt16 content_id = mapRoot->getDocument()->getElementNameIndex(U"content");
     lUInt16 text_id = mapRoot->getDocument()->getElementNameIndex(U"text");
     int nb_items = mapRoot->isElement() ? mapRoot->getChildCount() : 0;
-    for (size_t i=0; i<nb_items; i++) {
+    for (int i=0; i<nb_items; i++) {
         ldomNode * navPoint = mapRoot->getChildNode(i);
         if ( navPoint->getNodeId() != navPoint_id )
             continue;
@@ -177,7 +177,7 @@ void ReadEpubNcxPageList( ldomDocument * doc, ldomNode * mapRoot, LVPageMap * pa
     lUInt16 content_id = mapRoot->getDocument()->getElementNameIndex(U"content");
     lUInt16 text_id = mapRoot->getDocument()->getElementNameIndex(U"text");
     int nb_items = mapRoot->isElement() ? mapRoot->getChildCount() : 0;
-    for (size_t i=0; i<nb_items; i++) {
+    for (int i=0; i<nb_items; i++) {
         ldomNode * pageTarget = mapRoot->getChildNode(i);
         if ( pageTarget->getNodeId() != pageTarget_id )
             continue;
@@ -217,7 +217,7 @@ void ReadEpubNavToc( ldomDocument * doc, ldomNode * mapRoot, LVTocItem * baseToc
     lUInt16 a_id = mapRoot->getDocument()->getElementNameIndex(U"a");
     lUInt16 span_id = mapRoot->getDocument()->getElementNameIndex(U"span");
     int nb_items = mapRoot->isElement() ? mapRoot->getChildCount() : 0;
-    for (size_t i=0; i<nb_items; i++) {
+    for (int i=0; i<nb_items; i++) {
         ldomNode * li = mapRoot->getChildNode(i);
         if ( li->getNodeId() != li_id )
             continue;
@@ -282,7 +282,7 @@ void ReadEpubNavPageMap( ldomDocument * doc, ldomNode * mapRoot, LVPageMap * pag
     lUInt16 li_id = mapRoot->getDocument()->getElementNameIndex(U"li");
     lUInt16 a_id = mapRoot->getDocument()->getElementNameIndex(U"a");
     int nb_items = mapRoot->isElement() ? mapRoot->getChildCount() : 0;
-    for (size_t i=0; i<nb_items; i++) {
+    for (int i=0; i<nb_items; i++) {
         ldomNode * li = mapRoot->getChildNode(i);
         if ( li->getNodeId() != li_id )
             continue;
@@ -315,7 +315,7 @@ void ReadEpubAdobePageMap( ldomDocument * doc, ldomNode * mapRoot, LVPageMap * p
         return;
     lUInt16 page_id = mapRoot->getDocument()->getElementNameIndex(U"page");
     int nb_items = mapRoot->isElement() ? mapRoot->getChildCount() : 0;
-    for (size_t i=0; i<nb_items; i++) {
+    for (int i=0; i<nb_items; i++) {
         ldomNode * page = mapRoot->getChildNode(i);
         if ( page->getNodeId() != page_id )
             continue;
@@ -919,7 +919,7 @@ LVStreamRef GetEpubCoverpage(LVContainerRef arc)
         ldomNode * metadata = doc->nodeFromXPath(lString32("package/metadata"));
         int nb_metadata_items = (metadata && metadata->isElement()) ? metadata->getChildCount() : 0;
         lUInt16 meta_id = doc->getElementNameIndex(U"meta");
-        for (size_t i=0; i<nb_metadata_items; i++) {
+        for (int i=0; i<nb_metadata_items; i++) {
             ldomNode * item = metadata->getChildNode(i);
             if ( item->getNodeId() != meta_id )
                 continue;
@@ -936,7 +936,7 @@ LVStreamRef GetEpubCoverpage(LVContainerRef arc)
         ldomNode * manifest = doc->nodeFromXPath(lString32("package/manifest"));
         int nb_manifest_items = (manifest && manifest->isElement()) ? manifest->getChildCount() : 0;
         lUInt16 item_id = doc->getElementNameIndex(U"item");
-        for (size_t i=0; i<nb_manifest_items; i++) {
+        for (int i=0; i<nb_manifest_items; i++) {
             ldomNode * item = manifest->getChildNode(i);
             if ( item->getNodeId() != item_id )
                 continue;
@@ -1285,7 +1285,7 @@ bool ImportEpubDocument( LVStreamRef stream, ldomDocument * m_doc, LVDocViewCall
             // if we're going to load the book (and all its other metadata) from cache.
             // Iterate all package/metadata/identifier
             lUInt16 identifier_id = doc->getElementNameIndex(U"identifier");
-            for (size_t i=0; i<nb_metadata_items; i++) {
+            for (int i=0; i<nb_metadata_items; i++) {
                 ldomNode * item = metadata->getChildNode(i);
                 if ( item->getNodeId() != identifier_id )
                     continue;
@@ -1298,7 +1298,7 @@ bool ImportEpubDocument( LVStreamRef stream, ldomDocument * m_doc, LVDocViewCall
         }
         if ( decryptor->hasIdpfObfuscatedItems() && !idpfUniqueIdentifier.empty() ) {
             lUInt16 identifier_id = doc->getElementNameIndex(U"identifier");
-            for (size_t i=0; i<nb_metadata_items; i++) {
+            for (int i=0; i<nb_metadata_items; i++) {
                 ldomNode * item = metadata->getChildNode(i);
                 if ( item->getNodeId() != identifier_id )
                     continue;
@@ -1348,7 +1348,7 @@ bool ImportEpubDocument( LVStreamRef stream, ldomDocument * m_doc, LVDocViewCall
         lString32 authors;
         // Iterate all package/metadata/creator
         lUInt16 creator_id = doc->getElementNameIndex(U"creator");
-        for (size_t i=0; i<nb_metadata_items; i++) {
+        for (int i=0; i<nb_metadata_items; i++) {
             ldomNode * item = metadata->getChildNode(i);
             if ( item->getNodeId() != creator_id )
                 continue;
@@ -1368,7 +1368,7 @@ bool ImportEpubDocument( LVStreamRef stream, ldomDocument * m_doc, LVDocViewCall
         lString32 subjects;
         // Iterate all package/metadata/subject
         lUInt16 subject_id = doc->getElementNameIndex(U"subject");
-        for (size_t i=0; i<nb_metadata_items; i++) {
+        for (int i=0; i<nb_metadata_items; i++) {
             ldomNode * item = metadata->getChildNode(i);
             if ( item->getNodeId() != subject_id )
                 continue;
@@ -1418,7 +1418,7 @@ bool ImportEpubDocument( LVStreamRef stream, ldomDocument * m_doc, LVDocViewCall
         bool hasSeriesIdMeta = false;
         // Iterate all package/metadata/meta
         lUInt16 meta_id = doc->getElementNameIndex(U"meta");
-        for (size_t i=0; i<nb_metadata_items; i++) {
+        for (int i=0; i<nb_metadata_items; i++) {
             ldomNode * item = metadata->getChildNode(i);
             if ( item->getNodeId() != meta_id )
                 continue;
@@ -1470,7 +1470,7 @@ bool ImportEpubDocument( LVStreamRef stream, ldomDocument * m_doc, LVDocViewCall
             lString32 seriesId;
             // Iterate all package/metadata/meta
             lUInt16 meta_id = doc->getElementNameIndex(U"meta");
-            for (size_t i=0; i<nb_metadata_items; i++) {
+            for (int i=0; i<nb_metadata_items; i++) {
                 ldomNode * item = metadata->getChildNode(i);
                 if ( item->getNodeId() != meta_id )
                     continue;
@@ -1575,7 +1575,7 @@ bool ImportEpubDocument( LVStreamRef stream, ldomDocument * m_doc, LVDocViewCall
         CRLog::debug("opf: reading items");
         // Iterate all package/manifest/item
         lUInt16 item_id = doc->getElementNameIndex(U"item");
-        for (size_t i=0; i<nb_manifest_items; i++) {
+        for (int i=0; i<nb_manifest_items; i++) {
             if ( metadataOnly && !look_for_coverid && !look_for_epub3_cover && !look_for_coverxhtml_id) {
                 // No more stuff to find, stop iterating
                 break;
@@ -1719,7 +1719,7 @@ bool ImportEpubDocument( LVStreamRef stream, ldomDocument * m_doc, LVDocViewCall
                     pageMapHref = LVCombinePaths(codeBase, page_map->href);
                 // Iterate all package/spine/itemref (each will make a book fragment)
                 int nb_itemrefs = spine->getChildCount();
-                for (size_t i=0; i<nb_itemrefs; i++) {
+                for (int i=0; i<nb_itemrefs; i++) {
                     ldomNode * item = spine->getChildNode(i);
                     if ( item->getNodeId() != itemref_id )
                         continue;
